+++ AIPScannerConfig/functions/Start-PrerequisiteCheck.ps1
@@ -12,7 +12,7 @@
     .PARAMETER SqlRemote
         Switch to check for remote SQL instnaces
 
-    .PARAMETER SqlRemote
+    .PARAMETER UserDefinedSqlInstance
         User specified SQL Instance
 
     .PARAMETER Confirm
@@ -86,8 +86,6 @@
     process {
         if (-NOT (Assert-ElevatedPermission)) { return }
         Assert-IEEnhancedSC
-        $OriginalPreference = $ProgressPreference
-        $ProgressPreference = "SilentlyContinue"
         Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message2'
         $ServerVersion = Get-CimInstance Win32_OperatingSystem | Select-Object -ExpandProperty Caption
 
@@ -95,10 +93,10 @@
                 -or ($ServerVersion -like "Microsoft Windows Server 2016*")`
                 -or ($ServerVersion -like "Microsoft Windows Server 2019*")`
                 -or ($ServerVersion -like "Microsoft Windows Server 2022*")) {
-            Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message3'
+            Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message3' -StringValues $ServerVersion
         }
         else {
-            Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message4'
+            Write-PSFMessage -Level Host -String 'Start-PrerequisiteCheck.Message4'
             return
         }
 
@@ -126,13 +124,17 @@
 
         try {
             Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message11'
-            if (Test-NetConnection -InformationLevel Quiet) {
+            $OriginalPreference = $ProgressPreference
+            $ProgressPreference = "SilentlyContinue"
+            if (Test-NetConnection -ComputerName 'outlook.office365.com' -InformationLevel Quiet) {
                 Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message12'
             }
             else {
-                Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message13'
+                Write-PSFMessage -Level Host -String 'Start-PrerequisiteCheck.Message13'
                 return
             }
+            # Restore original preferences
+            $ProgressPreference = $OriginalPreference
 
             Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message14'
 
@@ -148,14 +150,14 @@
                 }
 
                 if (-NOT ($Instancefound)) {
-                    Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message16'
+                    Write-PSFMessage -Level Host -String 'Start-PrerequisiteCheck.Message16'
                     return
                 }
             }
             else {
                 # Check local user defined instance
                 if ($UserDefinedSqlInstance) {
-                    if (Get-SqlInstance -ServerInstance $UserDefinedSqlInstance -ErrorAction Continue) {
+                    if (Get-SqlInstance -ServerInstance $UserDefinedSqlInstance -ErrorAction SilentlyContinue) {
                         Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message17' -StringValues $UserDefinedSqlInstance
                     }
                     else {
@@ -164,11 +166,11 @@
                 }
 
                 # Check local default defined instance
-                if (Get-SqlInstance -ServerInstance ("$env:COMPUTERNAME\SQLEXPRESS") -ErrorAction Continue) {
+                if (Get-SqlInstance -ServerInstance "$env:COMPUTERNAME\SQLEXPRESS" -ErrorAction SilentlyContinue) {
                     Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message17' -StringValues "$env:COMPUTERNAME\SQLEXPRESS"
                 }
                 else {
-                    Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message18'
+                    Write-PSFMessage -Level Host -String 'Start-PrerequisiteCheck.Message18'
                     return
                 }
             }
@@ -179,6 +181,7 @@
             else {
                 Write-PSFMessage -Level Verbose -String 'Start-PrerequisiteCheck.Message20'
                 New-AIPSystemAccount
+                Add-AccountToSQLRole
             }
 
             New-AIPFileShare
@@ -188,9 +191,6 @@
             else {
                 return
             }
-
-            # Restore original preferences
-            $ProgressPreference = $OriginalPreference
         }
         catch {
             Stop-PSFFunction -String 'Start-PrerequisiteCheck.Message21' -EnableException $EnableException -Cmdlet $PSCmdlet -ErrorRecord $_